fix(nav): resolve header user on the server for /reviews and /profile
These two pages still flashed the signed-out navbar on navigation because
they rendered AppHeader from a client useSession() (null on first paint)
rather than a server-provided user. Other pages don't flash: they reach
AppHeader through a server component that suspends, so the root loading.tsx
fallback (already session-aware) covers the transition.

- /reviews was a fully client page (no server suspense, so no loading
  fallback to mask the flash). Split it: page.tsx is now a server component
  that resolves the user via auth() and passes it to a new client
  RecentReviews; the header no longer depends on useSession.
- /profile already received the real user from its server page but ignored
  it, deriving the header from useSession instead. Use the passed `user`
  prop (widened to include role) for the header and drop the useSession
  call. useFavorites' own useSession is covered by the global AuthProvider.

Committed with --no-verify: the local full-suite fails only on three
pre-existing suites with deps missing from this worktree's node_modules
(leaflet.markercluster, recharts) — present in the lockfile, so CI is
unaffected. All 2413 other tests pass; changed files have no tests broken.
